@@ -11,8 +11,19 @@ import sys; sys.path.append(str(Path(__file__).parent.parent))
 from acids_dataset import get_metadata_from_path, get_writer_class_from_path
 from acids_dataset import writers, transforms as adt
 from acids_dataset.features import AcidsDatasetFeature, ModuleEmbedding
-from acids_dataset.utils import GinEnv, parse_features, feature_from_gin_config, set_gin_constant
+from acids_dataset.utils import GinEnv, parse_features, transform_from_gin_config, set_gin_constant
 
+
+def import_flags():
+    flags.DEFINE_string('path', None, 'dataset path', required=True)
+    flags.DEFINE_string('module', None, 'path to module (.ts)', required=True)
+    flags.DEFINE_string('method', "forward", 'model method to call', required=False)
+    flags.DEFINE_string('name', None, 'embedding name (default : filename)', required=False)
+    flags.DEFINE_multi_string('transform', [], help='transforms for different embedding views.')
+    flags.DEFINE_integer('model_sr', None, 'original sample rate of target model', required=False)
+    flags.DEFINE_string('device', None, 'device id used for computation')
+    flags.DEFINE_boolean('overwrite', False, help="recomputes the feature if already present in the dataset, and overwrites existing files")
+    flags.DEFINE_boolean('check', True, help="recomputes the feature if already present in the dataset")
 
 def add_embedding_to_dataset(
     path, 
@@ -21,12 +32,20 @@ def add_embedding_to_dataset(
     module_sr: int | None = None,
     method: str = "forward",
     transforms: List[str | adt.Transform] = [],
-    embedding_name: str | None = None,
+    name: str | None = None,
     device: str | None = None,
     check: bool = False,
     overwrite: bool = False,
     ):
+
     path = Path(path)
+    if name is None: 
+        if module is None: 
+            name = name or Path(module_path).stem
+        elif module_path is None:
+            name = name or type(module).__name__.capitalize()
+        else:
+            raise ValueError('either module or module_path must be given')
 
     # parse gin constants
     gin.add_config_file_search_path(Path(__file__).parent / "configs")
@@ -36,34 +55,25 @@ def add_embedding_to_dataset(
     set_gin_constant('CHANNELS', metadata['channels'])
     set_gin_constant('DEVICE', device)
 
-    if ((embedding_name is None) and (module_path is not None)):
-        embedding_name = os.path.splitext(os.path.basename(module_path))[0]
+    operative_transforms = []
+    for t in transforms:
+        if isinstance(t, str):
+            operative_transforms.append(transform_from_gin_config(t))
+        elif isinstance(t, AcidsDatasetFeature):
+            operative_transforms.append(t)
+
+    print(transforms, operative_transforms)
+
 
     # parse module
     module_feature = ModuleEmbedding(module=module, 
                                      module_path=module_path, 
                                      module_sr=module_sr, 
                                      method=method, 
-                                     transforms=transforms, 
+                                     transforms=operative_transforms, 
                                      sr=metadata['sr'], 
-                                     name=embedding_name)
-
-    # parse contrastive transforms
-    operative_transforms = []
-    for i, f in enumerate(transforms):
-        if isinstance(f, str):
-            if os.path.splitext(f)[1] == "": f += ".gin"
-            if os.path.exists(f):
-                gin.add_config_file_search_path(Path(f).parent)
-            try:
-                gin.parse_config_file(f)
-            except TypeError as e:
-                print('[error] problem parsing configuration %s'%f)
-                raise e
-            with GinEnv(f):
-                operative_transforms.extend(feature_from_gin_config(f))
-        elif isinstance(f, AcidsDatasetFeature):
-            operative_transforms.append(f)
+                                     name=name, 
+                                     device=device)
 
     # build writer
     writer_class = get_writer_class_from_path(path)
@@ -76,13 +86,14 @@ def add_embedding_to_dataset(
 
 
 def main(argv):
+    FLAGS = flags.FLAGS
     add_embedding_to_dataset(
         FLAGS.path, 
         module_path=FLAGS.module, 
         module_sr=FLAGS.model_sr, 
         method=FLAGS.method, 
         transforms=FLAGS.transform, 
-        embedding_name=FLAGS.embedding_name,
+        name=FLAGS.name,
         device=FLAGS.device, 
         check=FLAGS.check, 
         overwrite=FLAGS.overwrite
@@ -90,14 +101,4 @@ def main(argv):
         
 
 if __name__ == "__main__":
-    FLAGS = flags.FLAGS
-    flags.DEFINE_string('path', None, 'dataset path', required=True)
-    flags.DEFINE_string('module', None, 'path to module (.ts)', required=True)
-    flags.DEFINE_string('method', "forward", 'model method to call', required=True)
-    flags.DEFINE_multi_string('transforms', [], help='transforms for different embedding views.')
-    flags.DEFINE_string('name', None, 'embedding name (default: model name)', required=True)
-    flags.DEFINE_integer('model_sr', None, 'original sample rate of target model', required=True)
-    flags.DEFINE_string('device', None, 'device id used for computation')
-    flags.DEFINE_boolean('overwrite', False, help="recomputes the feature if already present in the dataset, and overwrites existing files")
-    flags.DEFINE_boolean('check', True, help="recomputes the feature if already present in the dataset")
     app.run(main)
